Iguana: address review round 4
detail.h: reject the 0x80 NOP token explicitly. A block can be padded with them while still producing its declared output, so the per-command progress check does not catch it inside a single DecodeIguana command.

iguana.cc: the ANS input size assertion casts to uint64_t first, since size_t may be 32-bit, where the shift itself would be a warning or error.

iguana_test: cover both, with a control block that differs only by the NOPs and must decode, so the rejection is attributable to the token rather than to the declared length.

+++ hwy/contrib/iguana/iguana.cc
@@ -429,7 +429,8 @@ HWY_CONTRIB_DLLEXPORT std::vector<uint8_t> Compress(const uint8_t* data,
 
     for (size_t i = 0; i < kStreamCount; ++i) {
       // The ANS coder would wrap around (and divide by zero) at 4 GiB.
-      HWY_DASSERT(ustreams[i].size() < (size_t{1} << 32));
+      HWY_DASSERT(static_cast<uint64_t>(ustreams[i].size()) <
+                  (uint64_t{1} << 32));
       Bytes cs = Ans32Encode(ustreams[i].data(), ustreams[i].size());
       const double ratio = ustreams[i].empty()
                                ? 1e9
